+++ ICD.Connect.Routing.Extron/Devices/Endpoints/IDtpHdmiDeviceSettings.cs
@@ -1,5 +1,6 @@
 using ICD.Connect.Devices;
 using ICD.Connect.Routing.Extron.Devices.Switchers;
+using ICD.Connect.Routing.Extron.Devices.Switchers.DtpCrosspoint;
 using ICD.Connect.Settings.Attributes.SettingsProperties;
 
 namespace ICD.Connect.Routing.Extron.Devices.Endpoints