+++ ICD.Connect.Routing.Extron/Devices/Endpoints/Rx/DtpHdmiRx.cs
@@ -2,6 +2,7 @@
 using ICD.Connect.Protocol.Ports;
 using ICD.Connect.Protocol.Ports.ComPort;
 using ICD.Connect.Routing.Extron.Devices.Switchers;
+using ICD.Connect.Routing.Extron.Devices.Switchers.DtpCrosspoint;
 using ICD.Connect.Settings.Core;
 
 namespace ICD.Connect.Routing.Extron.Devices.Endpoints.Rx